Adding HomePage Component
Performed the migration of the App file code
to the index file of the home folder, the component
HomePage is called in the App provisionally.
